requirements.txt
os
requests
psycopg2-binary
python-dotenv
pandas
fastapi
pydantic
uvicorn
numpy